python: Move PackageLoader into dazl.protocols; make it fully async.
